BulletTracer collision check recursion fix
-  Added a limit to the number of times a BulletTracer's master check will recurse.

This fixes the targets in the INTERMAP training area causing the engine to hard lock.

When I added the Sherman, I set up BulletTracers to not collide with an actor that had the same master as the origin actor of the tracer; this check is recursive, so it finds the "top-level" master (so if the shooter has a master, and that master has a master, it will not collide with any 'descendants' of the master's master's master.  In some cases, this check can result in an infinite loop if the master's master's master (for example) is the original actor.  This change limits the looping to 10 iterations for each check.