C2M4: Fixed a few things spotted in MrIcarus' playthrough
C2M3: Started working on the eastern part of this map again. MrIcarus
playthrough was very useful in this department. The map is too long and
especially the underground area doesn't feel right, it's more redundant
and slows down the pace. I will completely rebuild that into a shorter
area.