AutoDrone behaviour overhaul
The missile attacks have been turned into melee attacks which are
activated within a range of 256 map units, which will prevent the player
from getting too close as they are very deadly.

The missile attack is now some kind of a mixture between the "Arch-Vile"
mechanic and the rail attack the drone had before. It aims at you for a
while (indicated by a red dot) in which time you can look for cover. If
you are not fast enough, you will get a railgun hit blast.

I think this mechanic works pretty nice. Already tried it in the boss
battle of C2M2 and it worked quite well. The fight is a challenge,
without being too hard or unfair, and it lasts for a while. Looking
forward to your thoughts friends :)