C3M5_C: Did some more work to the crash stuff
@AFADoomer, no idea but I think I broke the crash script somehow, check
line 94 for more clues about it. On top of that, after executing the
"EngineDestroy" script, I'd like to have the player visible but still in
his SS uniform. What is the right way to do that? And I also noticed
that the player movement is somehow quite strange when the tilt is
active.

@Ozymandias81 would it be possible to further improve the metal debris
("FallingMetal") actor to make it more random? I imagine falling little
pipes, falling little pieces of electronics, different metal pieces,
screws, all that kind of stuff. Also different sounds that are
randomized to make the actor not sound too generic.