Improve machinegun turret interactivity
Display turret ammo and heat when player is using the turret by using a proxy (TurretWeapon)
Change TurretGun shotcount to an int
Replace A_SpawnItemEx with Spawn so that the resulting actor can be cast to TurretGun
Re-structure TurretStand code: Move player's interactions to DoInteractions(), and add LeaveTurret() and GunIsIdle() methods. I made them virtual so that if another turret type is added (e.g. mortar), the code for the other turret can be adjusted accordingly.
Allow player to manually reload the turret by pressing the reload button.
Lock player to turret if it is being fired or reloaded

Also, I would like to change the player's height so he is slightly harder to hit while using the turret.