C3M5_C: Almost done, needs final tweaks by @AFADoomer
The map can be played from the beginning to the end. The timer at the
end is a bit short - instead of 120 seconds I'd like to have 180 seconds
but I have no idea how to adjust your timer @AFADoomer .

Regarding the scripts itself: As soon as the player picks up a key, a
script should be launched that makes certain enemies with a TID attack
him. Tried it with the red key but didn't get it to work.

Also when you are escaping and the airship has a certain pitch, jumping
makes you go ultra fast (check it after the final switch has been hit).

Also after the final switch is activated, not all enemies see the
player. No idea what I have done wrong.

I am sure you might find other things to improve, looking forward to
your feedback :)