C3M5_C: Various small fixes
- Fixed crosshair indicator in the C3M5_C end switch - for inventory items, the user_crosshair value takes a class name, not en editor ID number
- Tweaked tilted jump handling to factor in aircontrol value
  -- This fixes part of the jump problem, but...
  -- This still isn't perfect - there's still something wrong when you jump perpendicular to the angle of tilt, I think
- Added a line to the C3M5_C exit script to reset the level tilt on successful exit

I *think* that the blocking issues in the overhead area are being caused by line 179 of the SCRIPTS lump - setting blocking on lines tagged 181 to BLOCK_RAILING.  I've commented that line out in the version here, and the weird issues up top are fixed, but now the player can run off of the broken catwalk area, and there's nothing in place to either kill you or to allow you to get out.

Also, if there were this much fire inside of a hydrogen-filled balloon...  There would already have been a huge explosion.