Possible ladder script replacement
I've added a "ClimbableZone" actor that you place at the top of a texture-based ladder, then adjust the Y scale of the actor as appropriate until the zone is the height of the area that you want the player to be able to climb (it's invisible in game, but shows up as a translucent orange cuboid model in GZDB so you can easily adjust the size by sight).  PlayerFollowers can climb when near these actors as well as players.

I've added a TEST_LDR map as a quick proof of concept and demo.

It will be a lot of work to clear the old ladder script lines in every existing map and add ClimbableZone actors, but the effect is better, I think...

The new test map also shows off another actor type that I've been experimenting with...  RopeSpawner actors spawn a rope along the line of the actor's pitch on map load that is made up of many small model-based segments.  Vertical ropes are climbable using the same code as ClimbableZone actors, and horizontal ropes are walkable, like a tightrope.  I'm not sure what the performance hit of these might be in larger maps and in large numbers, but they could be a cool addition if used sparingly.

Miscellaneous other detailed documentation is at the top of ladder.txt...  These are still WIP, but final "enough" to show off.