Added knot model for rope
- Knots are automatically added at the origin of a rope and at the end of the rope for ropes that aren't vertical
- Also tweaked some code so that steep ropes are easier to navigate

TODO: I'm going to slightly re-work the rope spawning so that the placed object better and the in-game rope align properly (the rope spawns 4-8 map units below the placed actor right now).  Also going to add a "ropespawner destination" actor (probably just a mapspot that will be paired via TID with a rope spawner) so that you will be able to set up a rope by just placing a rope spawner and a destination, and giving them the same TID...  Mostly so you won't have to blindly aim the rope spawner in GZDB and hope that it hits in the right spot.