Various ZScript updates
- Minor tweaks to make horizontal ropes a little narrower to walk on
- Small tweak to player class to add a 10-second timeout to the medkit storage shelf wall use, so you can't just spam it
- Added a RopeSpawnerTarget actor that can be used to set the end point of a rope.

To use, just place a RopeSpawner or CableSpawner where you want the rope to start, then place a RopeSpawnerTarget where you want the rope to end.  Give both actors the same TID, and everything else will be handled automatically on map load.  If you don't set the TID, the actor will fall back to spawning the rope along whatever pitch you set it to have (the original behavior).

TEST_LDR has been updated with a demo - the walkable rope and endpoint both have a TID of 1, and the walkable cable and endpoint both have a TID of 2.