Fix C2M5_B mission number bug
I did a playthrough of WolfenDoom chapter 2, and after completing C2M5_B, I was sent all the way back to the first mission.

This commit fixes that bug, as well as provides a mechanism for forcing certain maps to certain mission numbers.

Now, mission numbers are the level number + a multiple of 100 above 100. Mission numbers above 100, but below 200 are secret missions.