C3M5_C: Final tweaks, should work well now except a few things I noticed
I populated the ship with more things to "happen", with more interactive
actions, making the whole ship "more alive" and worthwile being
discovered. I think I did a good job with it. Try it and let me know
what you think.

@AFADoomer - The "Toilet SS Officer" can't be made sneakable which makes
it disappear if it is set to through UDMF

@Talon1024 - For C3M0_A we implemented the "Notes" system with the
letters that we can place here and there. I would like to add these
notes to places in Chapter 1 and Chapter 2 (as well as Chapter 3 while
being at it) to enhance the story, give more information and details
about the story, the side-quests and other things. As you are the man
for languages and stuff, how do you think can we add this? Should I add
my ideas and you improve the language or would you like to add the notes
and letters with approporate story-telling on your own?

@Ozymandias81 - I think there is a good amount of textures/flats which
have no sound attached. Would you mind taking a closer look at the maps
and adding the missing ones to the TERRAIN definition? I also noticed
that the colored doors in C3M5_C have no brightmap, can you add them?