C3M1: Started improving the map in terms of gameplay after having it not touched for a while.
Something seems to break certain scenes when you are still in the
stealth mode at the beginning. The latest GZDoom crashes without a bug
report in numerous instances at the first third of the map. @AFADoomer
can you confirm this? I suppose it is directly connected to the stealth
system.