More improvements to ZScript compass
Allow actors to be added to the compass using the static Add method of BoACompass and the BoA_CompassQueue ACS script.
Use the RenderEvent variables instead of the player's position and angle to make the compass smooth.
Tweak the compass center to be a little bit higher.
Improve compassIcon method of CompassHandler so that the icon selection logic is more similar to the original ACS script.