Remove special from line 16876 on C3M1
Since line 16876 was triggered by the player pressing the use key, and ran a script to spawn a guard inside, I removed it, since I didn't think a lot of people would notice.

For reference, line 16876 is on the highest step to the rear entrance to the camp, where there is a guard you have to snipe from the tower in order to proceed.