Tracer Optimization
- Changed BulletTracer to only spawn a TracerTrail actor when it is first spawned.  Previously it spawned a new one every tic as it looped in the Spawn state.
- Decreased the speed of TracerTrail fadeout so that it is more realistic (they take 2-3 seconds to fade out completely)

With the old code, just one shot could end up spawning hundreds of TracerTrail actors (that all faded out in 4 tics) as it moved across the map...  Now that same shot only spawns 1 TracerTrail actor.

Prior to this change I was getting hard engine crashes in C1M4 in the initial fight on top of the dam (usually only during a full-on firefight with all enemies awake and firing at me).  Now I can play through with no problems.  Not sure if there's a garbage collection issue in the engine that we exposed, or what...  I'll try to narrow things down, but this seems to eliminate the trigger for me, at least.