Experimental PostProcessing shaders (by Nash)
I added the Post Processing shaders by @nashmuhandes as an experimental test. I didn't ask for permission yet, as this is only a first try, would like to know what you think about such features in general (good thing is: It can be activated and deactivated in the menue).

The "noise" and "vignette" could be very fitting in regards to the 1942'ish feeling we are approaching.

The "motion blur" (it isn't working yet, maybe I am missing something) would fit well in regards to the level of "realism" we are approaching (e.g. flinch).

I am still a bit sceptic about the "flare" one though, maybe I don't have the right settings as for now.

Let me know what you think.