ZScript GZDoom version 3.8 compatibility bump
- Now that we seem to have a stable dev build...  bumped minimum ZScript level to 3.8 and updated all now-deprecated commands as appropriate.

Note that his means that a recent dev build (~3.8 pre 226) or newer will be required in order to run BoA as of this commit!

TODO: Strobe light emitters are broken in recent builds after Graf did a re-write of how dynamic lights are handled internally (the strobes spam spawn lights over and over again in a circle now).  I haven't figured out an efficient way to fix them, yet.