Longinus Hitler Updates
- Significanlty reduced code duplication in Longinus Hitler by ZScriptifying the NaziBoss base class and adding some special handling.
- This fixes the sprite set not changing properly when the chaingun is dropped

Basically, the actor is automatically forced to use the sprite defined in the new "Wounded" state as soon as the 'wounded' boolean is set to true.  Since the frame order of the sprite sets is the same for most states (except for the Pain state, which has an exception coded in), this lets you seamlessly switch to the second sprite set once the Wound state is entered for the first time.