C3M4: Connected (almost) the tunnel with the city, applied Zyklon to the city
ZOMBIES: Made the ZyklonClouds non-additive and removed the gllight, so it doesn't get too vibrant when too many zombies are attacking.

ZYKLONMASK: Made the visor larger and the overlay less noticeable, otherwise you aren't able to even recognize my beautifully crafted zyklon town ;) @talon1024 - How hard would it be to add a subtle "breathing" sound to the gasmask and spacesuit?