Displacement maps addition
I added a few examples for the displacement maps on C1M1. It's a very subtle effect and I am pretty sure it isn't necessary on all textures but in my eyes it adds a special depth and realism to some textures. For the demo, check the sand, the rocks, brick and also the windows on the first houses on C1M1.