New pen holder added plus alarm problem
DECORATE, DEFS, SPRITES, CREDITS & MAPS
:new: Added new pen holder on C2M1 start and updated credits
:new: Attempt to fix Alarm issues adding ChangeTid to clear tids given
to alarms
:warning: Sneakable enemies can still activate them, at least changing
the tid doesn't rollback the AlarmPanel sprite as before: is it possible
to fix removing the interactivity on broken panels @AFADoomer ? I
noticed this issue at the very start of c2m1, thought it was because
Tids sharing but enemies can still activate destroyed panel... Should
this happen also to bells then? To test this simply break an alarm,
spawn a new enemy in front of it, alert it and you'll see it turn the
alarm again, even if it shouldn't be able anymore. Happens also with
non-spawned enemies, so just actors in place in a map. I am saying this
to see if there is a solution without editing each map with alarms...