C3M5_A: Finished the train sequence
Added a lot of material (sounds, textures) and finished the train sequence. Enter the train, hit the launch button, get back into passenger area, follow the train to the end, find a nice little surprise and escape the tunnel.

Through this final area, the player gets to the Tempelhof airport :)