C3M0_B: Set up the plans, need some work done before
For the second secret mission I have plans for a prequel to Blade of Agony that tells you what has happened to Ctp. James Ryan before he was captured. What is it all about?

After killing James Ryan in C3M3, you notice that his room is not under investigation anymore. You enter. On the table you find his mission briefing and a diary. After "using it" you get instantly moved to C3M0_B.

You start as Cpt. James Ryan (this is the moment where we need to use the monster sprites of Cpt. James Ryan (using FREExx psrites) and also a new complete set of mugshots for the player (based on MS_PRIS0) - sprite work by @doomjedi and @ozymandias81 , code work by @afadoomer (is that doable?).

The secret mission will actually tell the story of James Ryan's imprisonment and the reason what has happened to him and why he changed sides. This will give the mod and the character so much more depth. It's not as easy as it is with Hitler - Hitler is evil, but James Ryan got corrupted somehow (this is the story I am going to tell).

So, let me know what you think and let me know if you can prepare the stuff I need :)