More Keen work
- Added Princess Lindsey, Dopefish, Schoolfish, Inchworm, and Mimrock
- Added bubble sprites and sound for Dopefish burp
- Temp-fixed a very wrong sprite frame in the Mimrock
- Changed 'friendlies' - Council Member and Lindsey - to show up in new NPCs sub-folder in GZDB
- Tweaked Keen stunner ball to not work underwater
- Updated CM01 with all of the new actors

The Inchworms were used to launch an easter egg/secret level in the original game when you got them all grouped around you.  I've incorporated the same capability here, but it just prints a message right now...  Maybe a reward of some sort, or a way to get to another easter egg?