SM01: Final work, map is finished, ready for SM02 now!
@afadoomer - I think the Keen player class still needs some kind of screen indicator if he gets hurt. Otherwise you can get attacked from behind without even noticing until you die.

The red jumping ball should be used for getting over the spiked pit at 7426 | 2002 but unfortunately it doesn't move over the spikes. Can this be changed?

The vertical platorms "electricity" (movement indicator) doesn't rotate with the parent actor at 6976 | 1056.

The jumping blue enemies have a huge problems following you around the corners.

Btw, I would like to add the Berkeloid to the place where the Bird currently is but afaik it is not yet coded, is it?