C3M1: Finished the skybox and interconnection of the main building segments
This should heavily improve performance when climbing through the tower and also when moving in the outside area, as there is only one portal instead of two for the main building.

Will continue finishing up C3M1 now and clean up the remaining rooms, improve the gameplay and the map route.