C3M5_A NPC addition
Added a Wehrmacht soldier who hides in the sewers of the initial area of C3M5_A. The very first secret (with MP 40) on this map now has a switch behind the moving stone, which must be activated before it reaches the wall. The switch opens a grate in the tunnels, which is now a polyobject instead of a linedef. The new passage at (-18112, -12608, -200) leads to a dead end with a broken door, near which there is the new NPC. If talked to, before the last dialogue page he gives the player a dummy token; this token is only used once in a script for a door near (-13640, 10448, -8). It can be opened for supplies (possible design inconsistencies here).
Bugs: the name of NPC and his sprite are not shown, yet they do already exist in the repo. The conversation mugshots do not exist yet. The secret works as I intended though.