Improve GetInterceptPosition
Calculate squared length more efficiently by doing a dot product of the vector with itself.
Also, return the time it takes for a projectile to hit the target along with the intercept position. Note that GetInterceptPosition now returns 2 values instead of just one. If the intercept position cannot be calculated for whatever reason, return -1 as the time.