Attempt to fix a bug on C3M2
If the player does not approach or enter the ruined mansion from the front, Braunschweig, the Zyklon C tanks, and the flak cannons do not appear on the compass. This is an attempt to fix that. I marked all the lines around the ruined mansion to execute the "BF2_BRIEFING" script, and removed the "Actor enters sector" things, which are unreliable.