Camp Prisoner collision updates
- Partial reversion of 56c48fa5be9d04d185e8c8b17ec756783965ecdd.  You can't assign multiple species to an actor, so this broke the handling for player bullets passing through prisoners on lower skill levels.  THRUSPECIES also had the negative side effect of prisoners walking through the player and sometimes bunching up and walking inside of each other.

This commit makes the player pass through any PlayerFollower-species actor *if* the z height of the player is higher than that of the other actor.  So, effectively, you fall through the actor if you try to jump on top of them.  Only negative side effect is that you can also jump through the other actor while both at normal ground height if your timing is right, but that's less jarring than having them effectively not be solid when you're walking around.

Also adjusted scaling of prisoner actors, which has been coded in for 4 years, but never worked until I added a missing "Super.PostBeginPlay();" call to the Base class a few days ago.