C3M0_A: Brainstorming
@afadoomer - I am currently rethinking the gameplay in this map as I am not sure what the best decision is. That's why I want to ask you as well for your thoughts.

Currently we added the day-night cycle to the map, making it partially a really stresful job beating the map. No time to tell the story, no time to "discover" everything. On the other hand it's a nice feature in general to have the cycle. But is it really necessary?

I am thinking about stripping the whole day/night cycle from the map and turning it in a horror-survival with a lot of puzzles, a lot of story, resembling the gameplay of Resident Evil somehow.

What do you think of this decision? What would you suggest?