Handling for objects that can float in water
- Pushable crates will now float when pushed into water or other swimmable sector
- Oil Drums and Gas Drums will also float when they are placed in editor or otherwise spawned in water
- Also added Safe compass icon

I added this handling to the ModelBase base class, disabled by default, so most 3D map objects can be set up to use this effect.  Enable by using the ModelBase.Bouyancy property, set to a value that represents how much of the actor should stick out of the water when floating, ranged from 0.0 (disabled) to 1.0 (the entire actor sits on top of the water).  e.g., setting it to 0.2 leaves only the top 20% of the model sticking out of the water.  See TEST_LDR for demo.