C2M2: Did some fixes on the map regarding consistency and bugs
In general I have the feeling that this map needs some kind of overhaul (the same kind of overhaul that C1M4 got). It's very technical, extremely technical with little references to the other maps to be honest. I have no idea where to start though but I feel it's necessary. The route through the map needs to be more clear and in general it needs to fit the rest of the episode more. Let me know what you think though.

@afadoomer - The error I noticed with saving also happens in the INTERMAP when I "pukename mission x". It instantly crashes and I can't do any further testing in C2 atm.