C2M0_A: Did my first playtesting of the map, fixed a few things
and I have a request to @ozymandias81 and @talon1024 - I think the "zombie like" sounds for the Templars isn't appropriate. They are not dumb undead, they are knights of the Templar Order. I think something else need to be added here. No idea where to get that but I think some latin voiceovers would be cool here when seeing and attacking you.

Also I noticed at the waterfalls in the map that the "Watersplash" actors (special effects) make splashes. That shouldn't happen (you can hear them).

@afadoomer - Can you prepare some kind of an additional "papertype" and "font" in a more dirty and futuristic style for the "Papertexts" so I can also add them to Astrostein? I thought about some very rusty PDA (more the kind of Star Wars universe stuff, not the Star Trek one) with a fitting screen and screen font.