INTERMAP: Added some final text passages before and after Astrostein
With this being added, my story-revamp of C2 is finished. The next step is to improve C2M2 to make it fit better into the design canon of Blade of Agony. It will mostly involve design changes in terms of texture usage and structure implementation. I will try to get a similar result like in C1M4. And after that, I will continue with the story and playtesting of C3.