Alternative fix for rolling Nazis turning invisible when dying
With the way the fix for that bug was implemented by @AFADoomer, the enemies appeared to stand up in the middle of their dodge roll, even though they were still actually in the dodge roll, and it was weird because their hitbox wasn't as tall as them.

This should be a more "proper" fix for some rolling Nazis turning invisible when they die.