C3M1/C3INTRO
C3INTRO: I changed a few things in the introduction dream that I felt where out of place (Zombie files) and added things, that made sense when it comes to a dreaming sequence full of fears.

C3M1: Splitted the maps in two. I am not done with playtesting yet and the maps both need some adjustments but the main meat is done. Will continue on C3M1 tomorrow.