+STEN altfire
Added a semi-auto altfire mode to the STEN. Had to write some crutches (specifically NOAUTOFIRE and an A_Refire codepointer in the reload state), but the testing results seem quite good. The only side effect is that of the flag: now STEN acts like the rocket launcher or BFG9000 from Doom (it will not be fired if it is selected while fire is pressed); but this is minor since half of the weapons, even the Luger, have such behaviour (and STEN can now be considered a 'silent weapon').
Also buffed the STEN's silencer (reduced alert range by a quarter).
This commit is sponsored by RTCW.