Lifts, weapons, language
Hopefully finished with lifts. For some reason on C3M4 there were two lifts in different locations which shared the same tag :|
Astrostein grenade tokens are no longer taken with sv_infiniteinventory 1.
Added altfire for the chaingun (UMG43): you can hold (not just press!) the altfire button, and the rotor will spin without firing. This still alerts enemies. The benefit is that you do not have to wait for the barrels to spin up or down.
Updated the language file.