"Hack" the cyrillic font characters' palettes to look better
First, I converted the smallfont palette to a format usable by SLADE, and then I converted all of the cyrillic characters (and the dash) to paletted PNGs using SLADE and the smallfont palette, and then I used the palette hack to implant the palette and tRNS chunk, since SLADE always makes the palette and tRNS chunks 768 and 256 bytes long respectively.

This should make it so that the cyrillic font characters are more consistent with each other, so that some characters don't appear brighter or darker than others.