Crushes
TEXTURES
:new: Crushed recently added textures

:notebook: About the problem mentioned by @Tormentor667 on c3m5_b: I
don't have any issues but I guess it is related to spriteshadows embed
to the engine. For me it works even under 4.5-pre8, but anyway there are
lots of ambient smokes which slows the performance here and there. I use
OpenGL renderer so idk how it work on Vulkan.