C3M5_A/B/C: Improvements and additions to this set of maps
@ozymandias81 - Can you remove the messages that appear when you shoot "nazi cakes"? I wonder who has added these.

I am still far from being finished with C3M5_B but it's shaping up quite well. I am not sure if it makes sense to add puzzle stuff to the map, currently I think it works as is - action-packed and straight forward. We have puzzle in C3M5_C :)