Conversation Menu Music Volume Fix
- Force the music volume to the level's stored music volume when a conversation opens or closes.

Side effect of this change is that this effectively makes the conversation menus ignore the 'dlg_musicvolume' CVar (which isn't exposed in the menus anyway and is barely documented, so...).

This *should* fix the volume increase issues that were happening with conversations on INTERMAP before, though I could never fully recreate the problem.