Some work for C3INTRO
TEXTURES, CREDITS & MAPS
:new: Crushed recently added textures and updated credits
:new: C3INTRO - Clean the map a bit: https://imgur.com/a/4ucck4U

:notebook: I tried with no avail to create a checker using level_mapinfo
on boalib.acs to fix the foostep sounds issue on C3INTRO (falling into
the pyramid plays the step sound): my idea was once the script find out
we are on level 13 (c3intro), do not perform the check for
DoFootstepSounds... But I don't seem to be able to make it work,
probably because instead it is related to TERRAIN defs. Any idea anyone?