C1M6: Fixed the position of BJ at the epilogue to make it fit to the new scripts from @afadoomer
@afadoomer - There is another instance where the epilogue screen needs to be applied and that's script 113 in C2M0_B. I don't tend to be lazy but I am not sure if I get it how the script works, so I'd leave that implementation to you, is that okay?

Next to that one, we have 3 final places where we could take the advantage of that system: The cutscenes. We have CUTSCEN1, CUTSCEN2 and CUTSCEN3 that all show a quite similar text at their end just as in the epilogues. Do you think, the system could be implemented here as well for consistency?

In CUTSCEN1, the perfect spot would be the final scene with Hitler in front of the Eiffel Tower, in CUTSCEN2 the situation where Hitler examines the drone, in CUTSCEN3 in the Reichskanzlei during Hitler's speech.