INTERMAP: Added "debug area" where you can choose missions without using the console
The switch to activate the mode is the old chest position to the right (small switch at the brick border). Then a room opens that is at the back of the briefing room.

Currently I added C1M1-C1M6, maybe @ozymandias81 you can add the rest of the missions in the same sheme for me :)