C2M1: Added two different routes to the map (one shorter, one longer with the mayan artifact)
This hasn't been tested yet but the mechanics are there now and it should (or could) work (definitely needs fine-tuning, I will take care of that). Thanks to @afadoomer for the suggestion, I think that this is also a nice additional dynamic to the introduction map of chapter 2.  @talon1024 if you wish, feel free to improve the language for Darren message 110 - 112.