SM03: Fixed a misalignment of textures
@Talon1024 - I noticed a problem with the SM01-SM03 missions. Each time you return from them, all the scripts that are supposed to be executed before a new mission get executed again.

What does that mean? If you have talked to a few persons in the INTERMAP before entering the SM0x missions, as soon as you return, all the exclamation marks are active again. Can some kind of a check be implemented here if we return from a SM0x mission so the scripts don't run twice?