Improved the background for the dev-comments
Now they fit with the rest of the ingame graphics and the webpage. Also changed the standard name color to "gold" instead of other colors. Looks better on the background.

I guess it's time to add my own developer comments in my next playthrough ;) Still need to finish C3M0_A first... man this is really creating stress for me.