Make UberMutant have 2000 health by default
On C1M0, the UberMutant's health was 2000 because it was multiplied by 2. Since this makes the UberMutant's health greater than its default value, it was causing the health bar to show green, even after doing a significant amount of damage to it.

Also, remove UberMutant's health multiplier on C1M0.