Add a cutscene to the end of C3M5_B
Just a prototype for now. Not sure if I'll have time to polish it though... The main issue for now is the skybox.
I have used 'do not count towards completion' flag on enemy actors in the area, so they will not affect killcount (and I won't have to add their decorative variants :-) )