Fix trees becoming permanently invisible
If the EffectsManager culled a dormant tree, and spawned it back in, it would set the state label to "Inactive", which for some reason, caused the tree to go to state "Actor.0" instead of the first state in the respective actor's code. This change makes the Inactive state jump to the respective actor's Spawn state immediately.