Fixed some stuff mentioned on #188
DECORATE, ZSCRIPT & MAPS
:new: Tweaked a bit BatBase in order to have bats not blocking players
when fighting in narrow places, or even "ride" them on top... this since
applying Base.CANSQUISH does nothing for bats
:new: I was going to commit generalexplosion changes has requested by
Talon, but well I updated the repo a while ago and then pushed my
additional changes too for CTurrets
:new: C3M5_C - Tweaked skills for enemies present at redkey battle, now
2 officers appear only with skill 3-4
:new: C3M6_A - Fixed a couple of issues mentioned on #188

@AFADoomer is it possible to have FadeDeath for tanks? This way we can
have them fade on zones where players can get stuck, for example SSTank2
is very big and on C3M6_A there is that risk, but also on C1M0... I was
doing it myself, moving user_fadedeath and its property to Base and then
call CheckFadeDeath on DoDeath for tanks, but I doesn't seem to
understand how to transfer it to treads properly so idk the solution for
this. Any help?