Add pitch fix for 3D trees
If the player turns off 2D trees in the middle of a mission, far-away trees may be spawned in with the wrong pitch. This should fix that.
Also, nitpick order of code lines where an "effect" is culled, so that null is assigned to the EffectInfo before it is removed.