C3M5_A key graphics, C3M6_A zombie chamber
Added graphics for the surrendering soldier key.

The chamber still needs heavy balancing and style changes (I will take care of this a bit later). For some reason the NaziGeist spawners only produce one of them, I couldn't find out why. BTW I'd like to apply a Zyklon-styled effect to the instant-death floors, even better if with ghosts similar to those who spawn in barriers which can be opened with a number of souls — can it be done?

@Tormentor667 Could you please look at the general layout of the zombie chamber battle (i. e. run it with godmode) and tell me if it became better? If it totally did not, I will revert the changes and have to forget about it; if you decide that now it has potential, then I will continue adjusting its parameters.

The base.txt change fixes a division by zero which occured to me multiple times while testing.