C3M0_A: Some more tuning, much better when it comes to gameplay and story now
The map is finished except the remaining additional boss that I talked about with @DoomJedi and @AFADoomer - feel free to test and let me know what you find and what you would like to have balanced. The only thing I am going to add are some jump-scares to various places maybe but that's all for now.

@Ozymandias81 @DoomJedi I have a few suggestions, I hope you can help me with that:

- Can the opera singer also zombified/zyklon-c-fied? It doesn't make much sense to have a living singer there.
- Can Erik be given a more interesting attack like a Mancubis pattern when tossing knives? Also noticed full brightness when tossing knives, maybe removing that or adding a subtle brightmap would look better.
- Can Erik be given about 50 soul drops when he dies? Currently there is no use in killing him at all.