First try at improving the low-res skybox textures.
Everything works pretty well except some transitions from one side to another - to the top part of the skybox. I wonder if someone has an idea what to do here.

The "hires"-ones are in the apropriate folder, the original ones are still there in the textures/skybox folder.

Maybe @Talon1024 @Username-N00b-is-not-available you have idea how to make a good transition for these textures, as it might be worth it - especially if you take a look at the SUNSET_ ones - the lowres version are terrible but the hires ones are pretty convincing.