Add drivable Pz. Kpfw. IV
See also my recent checkbox addition to #280.
You can test the tank on TEST_LDR, or on any other map using 'summon PanzerIVTreads' or 'give PanzerIV'.
At the moment this is a mess of copy-pasted code, which can be improved if we abstract out some hardcoded actor classes in the drivable tank definition.