Remove level-specific items on all levels
I did a quick playtest of C3M3_A and B today, and I tried to look at the secret hint at the start of C3M3_B. Since it was referencing a sector in C3M3_A, I got a VM abort because of an out-of-bounds array access. I decided to re-do the code that clears level-specific items so that they would get cleared on every level, instead of just INTERMAP, so that this would not happen.