Playerfollower 'stay' command tweak
- Set up playerfollowers so that they can be configured to ignore your command to stay behind if you move outside of a specified range from them.

Currently set up so that if you command the PrisonerAgent to stay in place, the actor will only stop moving as long as you are within 4096 units of him.  Once you leave that radius, he returns to following your trail as normal, and will eventually catch up to you.

All other playerfollowers are unaffected by this change.